+++ Adj/Program/Primitive.hs
@@ -3,3 +3,5 @@ module Adj.Program.Primitive (module Exports) where
 import Adj.Program.Primitive.Generation as Exports
 import Adj.Program.Primitive.Maybe as Exports
 import Adj.Program.Primitive.Progress as Exports
+import Adj.Program.Primitive.Predicate as Exports
+import Adj.Program.Primitive.Boolean as Exports