+++ Adj/Program/Primitive/Predicate.hs
@@ -1,0 +1,6 @@
+module Adj.Program.Primitive.Predicate where
+
+import Adj.Algebra.Category (type (<--))
+import Adj.Program.Primitive.Boolean (Boolean)
+
+type Predicate = (<--) Boolean