+++ CHANGELOG.md
@@ -278,3 +278,4 @@
 * Rename `||` typeclass to `<?>`
 * Rename `resolver` method to `<?>`
 * Define `:.` family of type operators
+* Define `Adj.Program.Primitive.Predicate` module